+++ packages/charts/src/state/chart_state.ts
@@ -176,9 +176,11 @@ export const createChartStore = (chartId: string, title?: string, description?: 
         // TODO https://github.com/elastic/elastic-charts/issues/2078
         serializableCheck: false,
       }),
-    devTools: {
-      name: `@elastic/charts - ${chartId}`,
-    },
+    devTools: process.env.ELASTIC_CHARTS_DEV_MODE
+      ? {
+          name: `@elastic/charts - ${chartId}`,
+        }
+      : undefined,
   });
 };