+++ packages/charts/src/utils/themes/theme.ts
@@ -754,6 +754,18 @@ export interface RectStyle {
   widthRatio?: Ratio;
   /** applying textures to the bar on the theme/series */
   texture?: TexturedStyles;
+  /**
+   * The style applied to the rect when it is dimmed relative to other highlighted elements on the chart.
+   * This is typically used to visually de-emphasize the rect, for example, when another series is highlighted.
+   */
+  dimmed:
+    | { opacity: number }
+    | {
+        /** The fill color to use when the rect is dimmed. */
+        fill: Color | ColorVariant;
+        /** The opacity multiplier for the texture color when the rect is dimmed */
+        texture: { opacity: number };
+      };
 }
 
 /** @public */